search-service/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
faiss-cpu
joblib
python-dotenv
numpy

# python3 -m pip install -r requirements.txt